requirements.txt
streamlit
pandas
numpy
secure-smtplib
python-dotenv